@@ -33,10 +33,10 @@ import threading
 
 from storage import protocol
 from storage.shared.ecc import hash_data
+from storage.shared.subtensor import get_current_block
 from storage.validator.config import config, check_config, add_args
-from storage.validator.state import ttl_get_block, should_checkpoint
-from storage.validator.encryption import encrypt_data
-
+from storage.validator.state import should_checkpoint
+from storage.validator.encryption import encrypt_data, setup_encryption_wallet
 from storage.validator.store import store_broadband
 from storage.validator.retrieve import retrieve_broadband
 from storage.validator.network import (
@@ -118,12 +118,13 @@ class neuron:
         bt.logging.debug(f"wallet: {str(self.wallet)}")
 
         # Setup dummy wallet for encryption purposes. No password needed.
-        self.encryption_wallet = bt.wallet(
-            name=self.config.neuron.encryption_wallet_name,
-            hotkey=self.config.neuron.encryption_hotkey,
+        self.encryption_wallet = setup_encryption_wallet(
+            wallet_name=self.config.encryption.wallet_name,
+            wallet_hotkey=self.config.encryption.hotkey,
+            password=self.config.encryption.password,
         )
-        self.encryption_wallet.create_if_non_existent(coldkey_use_password=False)
         self.encryption_wallet.coldkey  # Unlock the coldkey.
+        bt.logging.info(f"loading encryption wallet {self.encryption_wallet}")
 
         # Init metagraph.
         bt.logging.debug("loading metagraph")
@@ -194,7 +195,7 @@ class neuron:
         if self.config.neuron.challenge_sample_size == 0:
             self.config.neuron.challenge_sample_size = self.metagraph.n
 
-        self.prev_step_block = ttl_get_block(self)
+        self.prev_step_block = get_current_block(self.subtensor)
 
         # Instantiate runners
         self.should_exit: bool = False
@@ -242,7 +243,11 @@ class neuron:
             - This property employs lazy loading and caching to efficiently manage the retrieval of top N validators.
             - The cache is updated based on specific conditions, such as crossing a checkpoint in the network.
         """
-        if self._top_n_validators == None or should_checkpoint(self):
+        if self._top_n_validators == None or should_checkpoint(
+            get_current_block(self.subtensor),
+            self.prev_step_block,
+            self.config.neuron.checkpoint_block_length,
+        ):
             self._top_n_validators = self.get_top_n_validators()
         return self._top_n_validators
 
@@ -369,7 +374,7 @@ class neuron:
 
         bt.logging.debug(f"returning user data: {decrypted_data[:100]}")
         bt.logging.debug(f"returning user payload: {user_encryption_payload}")
-        synapse.encrypted_data = base64.b64encode(user_encryption_payload)
+        synapse.encrypted_data = base64.b64encode(decrypted_data)
         synapse.encryption_payload = (
             json.dumps(user_encryption_payload)
             if isinstance(user_encryption_payload, dict)
@@ -452,6 +457,12 @@ class neuron:
         # In case of unforeseen errors, the API will log the error and continue operations.
         except Exception as e:
             bt.logging.error(traceback.format_exc())
+
+        # After all we have to ensure subtensor connection is closed properly
+        finally:
+            if hasattr(self, "subtensor"):
+                bt.logging.debug("Closing subtensor connection")
+                self.subtensor.close()
 
     def run_in_background_thread(self):
         """
