@@ -9,23 +9,29 @@ from storage.miner.set_weights import set_weights
 
 
 async def main(args):
-    subtensor = bt.subtensor(network=args.network)
-    wallet = bt.wallet(name=args.wallet, hotkey=args.hotkey)
-    metagraph = bt.metagraph(netuid=args.netuid, network=args.network, sync=False)
-    metagraph.sync(subtensor=subtensor)
-    my_subnet_uid = metagraph.hotkeys.index(wallet.hotkey.ss58_address)
+    try:
+        subtensor = bt.subtensor(network=args.network)
+        wallet = bt.wallet(name=args.wallet, hotkey=args.hotkey)
+        metagraph = bt.metagraph(netuid=args.netuid, network=args.network, sync=False)
+        metagraph.sync(subtensor=subtensor)
+        my_subnet_uid = metagraph.hotkeys.index(wallet.hotkey.ss58_address)
 
-    weights_were_set = set_weights(
-        subtensor=subtensor,
-        netuid=args.netuid,
-        uid=my_subnet_uid,
-        wallet=wallet,
-        metagraph=metagraph,
-        wandb_on=False,
-        wait_for_inclusion=False,
-        wait_for_finalization=True,
-    )
-    bt.logging.info(f"Were weights set? {weights_were_set}")
+        weights_were_set = set_weights(
+            subtensor=subtensor,
+            netuid=args.netuid,
+            uid=my_subnet_uid,
+            wallet=wallet,
+            metagraph=metagraph,
+            wandb_on=False,
+            wait_for_inclusion=False,
+            wait_for_finalization=True,
+        )
+        bt.logging.info(f"Were weights set? {weights_were_set}")
+    # After all we have to ensure subtensor connection is closed properly
+    finally:
+        if "subtensor" in locals():
+            subtensor.close()
+            bittensor.logging.debug("closing subtensor connection")
 
 
 if __name__ == "__main__":
