@@ -23,14 +23,14 @@ import aioredis
 import bittensor as bt
 
 
-async def store_chunk_metadata(r, chunk_hash, filepath, size, seed):
+async def store_chunk_metadata(r, chunk_hash, filepath, hotkey, size, seed):
     """
     Stores the metadata of a chunk in a Redis database.
 
     Args:
         r (redis.Redis): The Redis connection instance.
         chunk_hash (str): The unique hash identifying the chunk.
-        filepath (str): The file path associated with the chunk.
+        hotkey (str): Miner hotkey associated with the chunk.
         size (int): The size of the chunk.
         seed (str): The seed associated with the chunk.
 
@@ -39,6 +39,7 @@ async def store_chunk_metadata(r, chunk_hash, filepath, size, seed):
     # Ensure that all data are in the correct format
     metadata = {
         "filepath": filepath,
+        "hotkey": hotkey,
         "size": str(size),  # Convert size to string
         "seed": seed,  # Store seed directly
     }
@@ -48,14 +49,13 @@ async def store_chunk_metadata(r, chunk_hash, filepath, size, seed):
         await r.hset(chunk_hash, key, value)
 
 
-async def store_or_update_chunk_metadata(r, chunk_hash, filepath, size, seed):
+async def store_or_update_chunk_metadata(r, chunk_hash, filepath, hotkey, size, seed):
     """
     Stores or updates the metadata of a chunk in a Redis database.
 
     Args:
         r (redis.Redis): The Redis connection instance.
         chunk_hash (str): The unique hash identifying the chunk.
-        filepath (str): The file path associated with the chunk.
         size (int): The size of the chunk.
         seed (str): The seed associated with the chunk.
 
@@ -64,25 +64,27 @@ async def store_or_update_chunk_metadata(r, chunk_hash, filepath, size, seed):
     """
     if await r.exists(chunk_hash):
         # Update the existing entry with new seed information
-        await update_seed_info(r, chunk_hash, seed)
+        await update_seed_info(r, chunk_hash, hotkey, seed)
     else:
         # Add new entry
-        await store_chunk_metadata(r, chunk_hash, filepath, size, seed)
+        await store_chunk_metadata(r, chunk_hash, filepath, hotkey, size, seed)
 
 
-async def update_seed_info(r, chunk_hash, seed):
+async def update_seed_info(r, chunk_hash, hotkey, seed):
     """
     Updates the seed information for a specific chunk in the Redis database.
 
     Args:
         r (redis.Redis): The Redis connection instance.
         chunk_hash (str): The unique hash identifying the chunk.
+        hotkey (str): The caller hotkey value to be updated.
         seed (str): The new seed value to be updated.
 
     This function updates the seed information for the specified chunk hash.
     """
     # Update the existing seed information
     await r.hset(chunk_hash, "seed", seed)
+    await r.hset(chunk_hash, "hotkey", hotkey)
 
 
 async def get_chunk_metadata(r, chunk_hash):
@@ -141,8 +143,23 @@ async def get_total_storage_used(r):
 
 
 async def migrate_data_directory(r, new_base_directory, return_failures=False):
-    failed_filepaths = []
+    async for key in r.scan_iter("*"):
+        filepath = await r.hget(key, b"filepath")
+        old_base_directory = os.path.dirname(filepath.decode("utf-8"))
+        break
 
+    new_base_directory = os.path.expanduser(new_base_directory)
+    bt.logging.info(
+        f"Migrating filepaths for all hashes in Redis index from old base {old_base_directory} to new {new_base_directory}"
+    )
+
+    if not os.path.exists(new_base_directory):
+        bt.logging.info(
+            f"New base directory {new_base_directory} does not exist. Creating..."
+        )
+        os.makedirs(new_base_directory)
+
+    failed_filepaths = []
     async for key in r.scan_iter("*"):
         filepath = await r.hget(key, b"filepath")
 
@@ -152,12 +169,26 @@ async def migrate_data_directory(r, new_base_directory, return_failures=False):
             new_filepath = os.path.join(new_base_directory, data_hash)
 
             if not os.path.exists(new_filepath):
-                bt.logging.debug(
+                bt.logging.trace(
                     f"Data does not exist in new path {new_filepath}. Skipping..."
                 )
                 failed_filepaths.append(new_filepath)
                 continue
 
             await r.hset(key, "filepath", new_filepath)
+
+    if len(failed_filepaths):
+        if not os.path.exists("logs"):
+            os.makedirs("logs")
+        with open("logs/failed_filepaths.json", "w") as f:
+            json.dump(failed_filepaths, f)
+        bt.logging.error(
+            f"Failed to migrate {len(failed_filepaths)} files. These were skipped and may need to be migrated manually."
+        )
+        bt.logging.error(
+            f"Please see {os.path.abspath('logs/failed_migration_filepaths.json')} for a complete list of failed filepaths."
+        )
+    else:
+        bt.logging.success("Successfully migrated all filepaths.")
 
     return failed_filepaths if return_failures else None
