@@ -19,8 +19,91 @@ import time
 import wandb
 import bittensor as bt
 import traceback
-from .set_weights import set_weights, should_wait_to_set_weights
+from substrateinterface import SubstrateInterface
+from scalecodec import ScaleBytes
 from .utils import update_storage_stats
+from copy import deepcopy
+
+tagged_tx_queue_registry = {
+    "types": {
+        "TransactionTag": "Vec<u8>",
+        "TransactionPriority": "u64",
+        "TransactionLongevity": "u64",
+        "ValidTransaction": {
+            "type": "struct",
+            "type_mapping": [
+                ["priority", "TransactionPriority"],
+                ["requires", "Vec<TransactionTag>"],
+                ["provides", "Vec<TransactionTag>"],
+                ["longevity", "TransactionLongevity"],
+                ["propagate", "bool"],
+            ],
+        },
+        "TransactionValidity": "Result<ValidTransaction, TransactionValidityError>",
+        "TransactionSource": {
+            "type": "enum",
+            "value_list": ["InBlock", "Local", "External"],
+        },
+    },
+    "runtime_api": {
+        "TaggedTransactionQueue": {
+            "methods": {
+                "validate_transaction": {
+                    "params": [
+                        {
+                            "name": "source",
+                            "type": "TransactionSource",
+                        },
+                        {
+                            "name": "tx",
+                            "type": "Extrinsic",
+                        },
+                        {"name": "block_hash", "type": "Hash"},
+                    ],
+                    "type": "TransactionValidity",
+                },
+            },
+        }
+    },
+}
+
+
+def runtime_call(
+    substrate: SubstrateInterface, api: str, method: str, params: list, block_hash: str
+):
+    substrate.runtime_config.update_type_registry(tagged_tx_queue_registry)
+    runtime_call_def = substrate.runtime_config.type_registry["runtime_api"][api][
+        "methods"
+    ][method]
+    runtime_api_types = substrate.runtime_config.type_registry["runtime_api"][api].get(
+        "types", {}
+    )
+
+    # Encode params
+    param_data = ScaleBytes(bytes())
+    for idx, param in enumerate(runtime_call_def["params"]):
+        scale_obj = substrate.runtime_config.create_scale_object(param["type"])
+        if type(params) is list:
+            param_data += scale_obj.encode(params[idx])
+        else:
+            if param["name"] not in params:
+                raise ValueError(f"Runtime Call param '{param['name']}' is missing")
+
+            param_data += scale_obj.encode(params[param["name"]])
+
+    # RPC request
+    result_data = substrate.rpc_request(
+        "state_call", [f"{api}_{method}", str(param_data), block_hash]
+    )
+
+    # Decode result
+    result_obj = substrate.runtime_config.create_scale_object(runtime_call_def["type"])
+    result_obj.decode(
+        ScaleBytes(result_data["result"]),
+        check_remaining=substrate.config.get("strict_scale_decode"),
+    )
+
+    return result_obj
 
 
 def run(self):
@@ -48,134 +131,128 @@ def run(self):
         KeyboardInterrupt: If the miner is stopped by a manual interruption.
         Exception: For unforeseen errors during the miner's operation, which are logged for diagnosis.
     """
+    block_handler_substrate = SubstrateInterface(
+        ss58_format=bt.__ss58_format__,
+        use_remote_preset=True,
+        url=self.subtensor.chain_endpoint,
+        type_registry=bt.__type_registry__,
+    )
+
+    netuid = self.config.netuid
+
     # --- Check for registration.
     if not self.subtensor.is_hotkey_registered(
-        netuid=self.config.netuid,
+        netuid=netuid,
         hotkey_ss58=self.wallet.hotkey.ss58_address,
     ):
         bt.logging.error(
-            f"Wallet: {self.wallet} is not registered on netuid {self.config.netuid}"
+            f"Wallet: {self.wallet} is not registered on netuid {netuid}"
             f"Please register the hotkey using `btcli subnets register` before trying again"
         )
         exit()
 
-    # --- Run until should_exit = True.
-    self.last_epoch_block = self.metagraph.last_update[self.my_subnet_uid].item()
-    bt.logging.info(f"Miner starting at block: {self.last_epoch_block}")
+    tempo = block_handler_substrate.query(
+        module="SubtensorModule", storage_function="Tempo", params=[netuid]
+    ).value
 
-    # This loop maintains the miner's operations until intentionally stopped.
-    bt.logging.info(f"Starting main loop")
-    step = 0
-    try:
-        while not self.should_exit:
-            start_epoch = time.time()
+    last_extrinsic_hash = None
+    checked_extrinsics_count = 0
+    should_retry = False
 
-            # --- Wait until next epoch.
-            self.current_block = self.subtensor.get_current_block()
-            self.metagraph.sync(subtensor=self.subtensor)
-            self.last_epoch_block = self.metagraph.last_update[
-                self.my_subnet_uid
-            ].item()
+    def handler(obj, update_nr, subscription_id):
+        current_block = obj["header"]["number"]
+        block_hash = block_handler_substrate.get_block_hash(current_block)
+        bt.logging.debug(f"New block #{current_block}")
 
-            hyperparameters = self.subtensor.get_subnet_hyperparameters(
-                self.config.netuid, self.current_block
+        bt.logging.debug(
+            f"Blocks since epoch: {(current_block + netuid + 1) % (tempo + 1)}"
+        )
+
+        nonlocal last_extrinsic_hash
+        nonlocal checked_extrinsics_count
+        nonlocal should_retry
+
+        if last_extrinsic_hash != None:
+            try:
+                receipt = block_handler_substrate.retrieve_extrinsic_by_hash(
+                    block_hash, last_extrinsic_hash
+                )
+                bt.logging.debug(
+                    f"Last set-weights call: {'Success' if receipt.is_success else format('Failure, reason: %s', receipt.error_message['name'] if receipt.error_message != None else 'nil')}"
+                )
+
+                should_retry = False
+                last_extrinsic_hash = None
+                checked_extrinsics_count = 0
+            except Exception as e:
+                checked_extrinsics_count += 1
+                bt.logging.debug(f"An error occurred, extrinsic not found in block.")
+            finally:
+                if checked_extrinsics_count >= 20:
+                    should_retry = True
+                    last_extrinsic_hash = None
+                    checked_extrinsics_count = 0
+
+        if ((current_block + netuid + 1) % (tempo + 1) == 0) or should_retry:
+            bt.logging.info(
+                f"New epoch started, setting weights at block {current_block}"
             )
-            tempo = hyperparameters.tempo
-            bt.logging.info(f"Tempo for subnet: {tempo}")
+            with self.subtensor.substrate as substrate:
+                call = substrate.compose_call(
+                    call_module="SubtensorModule",
+                    call_function="set_weights",
+                    call_params={
+                        "dests": [self.my_subnet_uid],
+                        "weights": [65535],
+                        "netuid": netuid,
+                        "version_key": 1,
+                    },
+                )
 
-            # --- To control messages without changing time.sleep within the while-loop
-            # we can increase/decrease 'seconds_waiting_in_loop' without problems
-            # with 'seconds_to_wait_to_log_presence_message' we control the logging factor in the wait
-            seconds_waiting_in_loop = 1
-            presence_message_seconds_count = 0
+                # Period dictates how long the extrinsic will stay as part of waiting pool
+                extrinsic = substrate.create_signed_extrinsic(
+                    call=call, keypair=self.wallet.hotkey, era={"period": 1000}
+                )
 
-            while should_wait_to_set_weights(
-                self.current_block,
-                self.last_epoch_block,
-                tempo,
-            ):
-                # --- Wait for next bloc.
-                time.sleep(seconds_waiting_in_loop)
-                presence_message_seconds_count += seconds_waiting_in_loop
-                self.current_block = self.subtensor.get_current_block()
+                dry_run = runtime_call(
+                    substrate=substrate,
+                    api="TaggedTransactionQueue",
+                    method="validate_transaction",
+                    params=["InBlock", extrinsic, block_hash],
+                    block_hash=block_hash,
+                )
+                bt.logging.debug(dry_run)
 
-                if (
-                    presence_message_seconds_count
-                    % self.config.miner.seconds_to_wait_to_log_presence_message
-                    == 0
-                ):
-                    bt.logging.info(
-                        f"Miner UID {self.my_subnet_uid} running at block {self.current_block}..."
+                response = substrate.submit_extrinsic(
+                    extrinsic,
+                    wait_for_inclusion=False,
+                    wait_for_finalization=False,
+                )
+
+                result_data = substrate.rpc_request("author_pendingExtrinsics", [])
+                for extrinsic_data in result_data["result"]:
+                    extrinsic = substrate.runtime_config.create_scale_object(
+                        "Extrinsic", metadata=substrate.metadata
+                    )
+                    extrinsic.decode(
+                        ScaleBytes(extrinsic_data),
+                        check_remaining=substrate.config.get("strict_scale_decode"),
                     )
 
-                # --- Check if we should exit.
-                if self.should_exit:
-                    break
+                    if extrinsic.value["extrinsic_hash"] == response.extrinsic_hash:
+                        bt.logging.debug(
+                            "Weights transaction is in the pending transaction pool"
+                        )
 
-            presence_message_seconds_count = 0
-
-            # --- Set weights.
-            weights_were_set = False
-            if not self.config.miner.no_set_weights:
-                bt.logging.info(f"Setting weights on chain.")
-                # if both 'wait_for_*' args are False, weights_were_set = True
-                # even if they are not set yet or the extrinsic has failed
-                weights_were_set = set_weights(
-                    self.subtensor,
-                    self.config.netuid,
-                    self.my_subnet_uid,
-                    self.wallet,
-                    self.metagraph,
-                    self.config.wandb.on,
-                    tempo=tempo,
-                    wait_for_inclusion=self.config.miner.set_weights_wait_for_inclusion,
-                    wait_for_finalization=self.config.miner.set_weights_wait_for_finalization,
-                )
-            step += 1
-
-            # --- Update the metagraph with the latest network state.
-            if weights_were_set:
-                current_block = self.subtensor.get_current_block()
-                self.last_epoch_block = current_block
-                self.current_block = current_block
-
-                self.metagraph = self.subtensor.metagraph(
-                    netuid=self.config.netuid,
-                    lite=True,
-                    block=self.last_epoch_block,
-                )
-                log = (
-                    f"Step:{step} | "
-                    f"Block:{self.metagraph.block.item()} | "
-                    f"Stake:{self.metagraph.S[self.my_subnet_uid]} | "
-                    f"Rank:{self.metagraph.R[self.my_subnet_uid]} | "
-                    f"Trust:{self.metagraph.T[self.my_subnet_uid]} | "
-                    f"Consensus:{self.metagraph.C[self.my_subnet_uid] } | "
-                    f"Incentive:{self.metagraph.I[self.my_subnet_uid]} | "
-                    f"Emission:{self.metagraph.E[self.my_subnet_uid]}"
-                )
-                bt.logging.info(log)
-                if self.config.wandb.on:
-                    wandb.log(log)
-            else:
-                self.current_block = self.subtensor.get_current_block()
-                num_blocks_to_wait = 1
-                bt.logging.info(
-                    f"Weights were not set. Waiting {num_blocks_to_wait} blocks to set weights again."
-                )
-                time.sleep(
-                    num_blocks_to_wait * 12
-                )  # It takes 12 secs to generate a block
+                last_extrinsic_hash = response.extrinsic_hash
+                should_retry = False
 
             # --- Update the miner storage information periodically.
-            update_storage_stats(self)
+            if not should_retry:
+                update_storage_stats(self)
+                bt.logging.debug("Storage statistics updated...")
 
-    # If someone intentionally stops the miner, it'll safely terminate operations.
-    except KeyboardInterrupt:
-        self.axon.stop()
-        bt.logging.success("Miner killed by keyboard interrupt.")
-        exit()
+            if self.should_exit:
+                return True
 
-    # In case of unforeseen errors, the miner will log the error and continue operations.
-    except Exception as e:
-        bt.logging.error(traceback.format_exc())
+    block_handler_substrate.subscribe_block_headers(handler)
