@@ -20,53 +20,18 @@
 import torch
 import wandb
 import copy
-import time
 import math
 import hashlib as rpccheckhealth
 
-from math import floor
 from loguru import logger
 from dataclasses import asdict
 from typing import Callable, Any
-from functools import lru_cache, update_wrapper
 
 import storage
 import storage.validator as validator
 from storage.validator.event import EventSchema
 
 import bittensor as bt
-
-
-# LRU Cache with TTL
-def ttl_cache(maxsize: int = 128, typed: bool = False, ttl: int = -1):
-    if ttl <= 0:
-        ttl = 65536
-    hash_gen = _ttl_hash_gen(ttl)
-
-    def wrapper(func: Callable) -> Callable:
-        @lru_cache(maxsize, typed)
-        def ttl_func(ttl_hash, *args, **kwargs):
-            return func(*args, **kwargs)
-
-        def wrapped(*args, **kwargs) -> Any:
-            th = next(hash_gen)
-            return ttl_func(th, *args, **kwargs)
-
-        return update_wrapper(wrapped, func)
-
-    return wrapper
-
-
-def _ttl_hash_gen(seconds: int):
-    start_time = time.time()
-    while True:
-        yield floor((time.time() - start_time) / seconds)
-
-
-# 12 seconds updating block.
-@ttl_cache(maxsize=1, ttl=12)
-def ttl_get_block(self) -> int:
-    return self.subtensor.get_current_block()
 
 
 def should_reinit_wandb(self):
@@ -123,16 +88,9 @@ def reinit_wandb(self):
     init_wandb(self, reinit=True)
 
 
-def should_checkpoint(self):
+def should_checkpoint(current_block, prev_step_block, checkpoint_block_length):
     # Check if enough epoch blocks have elapsed since the last checkpoint.
-    a = ttl_get_block(self) % self.config.neuron.checkpoint_block_length
-    b = self.prev_step_block % self.config.neuron.checkpoint_block_length
-    bt.logging.debug(
-        f"should_checkpoint() calculation:\nblock % checkpoint_block_length: {ttl_get_block(self)} % {self.config.neuron.checkpoint_block_length}\n"
-        f"prev_step_block % checkpoint_block_length: {self.prev_step_block} % {self.config.neuron.checkpoint_block_length}\n"
-        f"{a} < {b}: {a < b}"
-    )
-    return a < b
+    return current_block - prev_step_block >= checkpoint_block_length
 
 
 def checkpoint(self):
